[#4] feat: Apply and Quit managed magazine base
Set basic flow of apply and quit magazine.
But sending mail to manager is not developed.
Magazine has username of manager, but not email address.
So account controller provide email by api.
But it does not work by rest-template from magazine package.
It could be because of connection pool problem.
I need to try using web-client, or associating magazine with account.